fix: handle data component divide
- DocumentListPage에서 data를 무한스크롤로 처리
- 그 data를 map컴포넌트에도 활용하여 마커 여러개로 처리
- DocsList에서 data를 prop 받아서 처리
