+++ src/components/common/document/DocsList.jsx
@@ -1,9 +1,5 @@
-import { useInfiniteQuery } from "@tanstack/react-query";
-import { useState, useRef, useEffect } from "react";
+import { useState } from "react";
 import DocsItem from "./DocsItem";
-import DocumentPage from "../../../pages/DocumentPage";
-import Loader from "../layout/Loader";
-import { docsList } from "../../../services/document";
 import { useNavigate } from "react-router-dom";
 import routes from "../../../routes";
 import styled from "styled-components";
@@ -20,48 +16,8 @@ const Container = styled.div`
   box-shadow: 10px 0px 5px 0px rgba(0, 0, 0, 0.106);
 `;
 
-const DocsList = () => {
+const DocsList = ({ data }) => {
   const navigate = useNavigate();
-
-  // 무한스크롤
-  const bottomObserver = useRef(null);
-
-  const { data, isLoading, isError, fetchNextPage, hasNextPage } =
-    useInfiniteQuery(
-      ["docs_list"],
-      ({ pageParam = 0 }) => docsList(pageParam),
-      {
-        getNextPageParam: (currentPage, allPages) => {
-          const nextPage = allPages.length;
-          return nextPage > 3 ? null : nextPage;
-        },
-      }
-    );
-
-  useEffect(() => {
-    const io = new IntersectionObserver(
-      (entries) => {
-        entries.forEach((entry) => {
-          if (entry.isIntersecting && !isLoading && hasNextPage) {
-            fetchNextPage();
-          }
-        });
-      },
-      {
-        threshold: 1.0,
-      }
-    );
-
-    if (bottomObserver.current) {
-      io.observe(bottomObserver.current);
-    }
-
-    return () => {
-      if (bottomObserver.current) {
-        io.unobserve(bottomObserver.current);
-      }
-    };
-  }, [isLoading, hasNextPage, fetchNextPage]);
 
   const docsData = data?.pages.flatMap((x) => x.data.response);
   const docsListArray = docsData || [];
@@ -86,8 +42,6 @@ const DocsList = () => {
     return (
       <>
         <Container>
-          {isLoading && <Loader />}
-          {/* {isError && <div>error</div>} */}
           {docsListArray.map((el) => (
             <DocsItem
               key={el.docsId}
@@ -96,8 +50,6 @@ const DocsList = () => {
               onClick={() => handleOnClick(el)}
             />
           ))}
-          <div style={{ height: "50px" }} ref={bottomObserver}></div>
-          {isLoading && !hasNextPage && <Loader />}
         </Container>
       </>
     );
